requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0

# Optional dependencies
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
